+++ controllers/networktopology.go
@@ -6,26 +6,29 @@ import (
 	"time"
 
 	"github.com/giantswarm/microerror"
-	corev1 "k8s.io/api/core/v1"
 	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/util/annotations"
+	capiconditions "sigs.k8s.io/cluster-api/util/conditions"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/giantswarm/aws-network-topology-operator/pkg/registrar"
+	nettopAnnotations "github.com/giantswarm/aws-network-topology-operator/pkg/util/annotations"
 )
 
-const FinalizerNetTop = "network-topology.finalizers.giantswarm.io"
+const (
+	FinalizerNetTop                             = "network-topology.finalizers.giantswarm.io"
+	networkTopologyCondition capi.ConditionType = "NetworkTopologyReady"
+)
 
 //counterfeiter:generate . ClusterClient
 type ClusterClient interface {
 	Get(context.Context, types.NamespacedName) (*capi.Cluster, error)
 	AddFinalizer(context.Context, *capi.Cluster, string) error
 	RemoveFinalizer(context.Context, *capi.Cluster, string) error
-	HasStatusCondition(ctx context.Context, cluster *capi.Cluster) bool
-	UpdateStatusCondition(ctx context.Context, cluster *capi.Cluster, status corev1.ConditionStatus) error
+	UpdateStatus(ctx context.Context, cluster *capi.Cluster) error
 }
 
 //counterfeiter:generate . Registrar
@@ -72,9 +75,10 @@ func (r *NetworkTopologyReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 		return ctrl.Result{}, nil
 	}
 
-	if !r.client.HasStatusCondition(ctx, cluster) {
+	if !capiconditions.Has(cluster, networkTopologyCondition) {
+		capiconditions.MarkFalse(cluster, networkTopologyCondition, "InProgress", capi.ConditionSeverityInfo, "")
 		// We're ok to continue if this fails
-		_ = r.client.UpdateStatusCondition(ctx, cluster, corev1.ConditionFalse)
+		_ = r.client.UpdateStatus(ctx, cluster)
 	}
 
 	if !cluster.DeletionTimestamp.IsZero() {
@@ -91,14 +95,21 @@ func (r *NetworkTopologyReconciler) reconcileNormal(ctx context.Context, cluster
 		return ctrl.Result{}, microerror.Mask(err)
 	}
 
+	defer func() {
+		_ = r.client.UpdateStatus(ctx, cluster)
+	}()
+
 	for _, reg := range r.registrars {
 		err = reg.Register(ctx, cluster)
 		if err != nil {
 			if errors.Is(err, &registrar.ModeNotSupportedError{}) {
+				capiconditions.MarkFalse(cluster, networkTopologyCondition, "ModeNotSupported", capi.ConditionSeverityError, "The provided mode '%s' is not supported", nettopAnnotations.GetAnnotation(cluster, nettopAnnotations.NetworkTopologyModeAnnotation))
 				return ctrl.Result{Requeue: false}, nil
 			} else if errors.Is(err, &registrar.TransitGatewayNotAvailableError{}) {
+				capiconditions.MarkFalse(cluster, networkTopologyCondition, "TransitGatewayNotAvailable", capi.ConditionSeverityWarning, "The transit gateway is not yet available for attachment")
 				return ctrl.Result{Requeue: true, RequeueAfter: time.Minute * 1}, nil
 			} else if errors.Is(err, &registrar.VPCNotReadyError{}) {
+				capiconditions.MarkFalse(cluster, networkTopologyCondition, "VPCNotReady", capi.ConditionSeverityInfo, "The cluster's VPC is not yet ready")
 				return ctrl.Result{Requeue: true, RequeueAfter: time.Minute * 1}, nil
 			}
 
@@ -106,9 +117,7 @@ func (r *NetworkTopologyReconciler) reconcileNormal(ctx context.Context, cluster
 		}
 	}
 
-	// We're ok to continue if this fails
-	_ = r.client.UpdateStatusCondition(ctx, cluster, corev1.ConditionTrue)
-
+	capiconditions.MarkTrue(cluster, networkTopologyCondition)
 	return ctrl.Result{Requeue: true, RequeueAfter: time.Minute * 10}, nil
 }