+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- updating the status condition to "true" when reconcile finished
+
 ## [1.1.3] - 2022-09-23
 
 ### Fixed